% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo.R
\name{fetch_player_data}
\alias{fetch_player_data}
\title{Fetch Player Data}
\usage{
fetch_player_data(playerid, matchtype = c("Test", "ODI", "T20"))
}
\arguments{
\item{playerid}{The player ID as given in Cricinfo}

\item{matchtype}{Which type of cricket matches do you want? Tests, ODIs or T20s?}
}
\description{
Fetch individual player data from all matches played. The function will scrape
the data from cricinfo and return a tibble with one line per innings for all
games a player has played.
}
\examples{
ElyssePerry <- fetch_player_data(275487,"T20")
MegLanning <- fetch_player_data(329336,"ODI")
SteveSmith <- fetch_player_data(267192,"Test")
library(ggplot2)
ggplot(MegLanning) + geom_point(aes(x=Date, y=Score, col=NotOut)) +
  ggtitle("Meg Lanning ODI Scores")

}
