% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo.R
\name{fetch_player_data}
\alias{fetch_player_data}
\title{Fetch Player Data}
\usage{
fetch_player_data(playerid, matchtype = c("Test", "ODI", "T20"))
}
\arguments{
\item{playerid}{The player ID as given in Cricinfo}

\item{matchtype}{Which type of cricket matches do you want? Tests, ODIs or T20s?}
}
\description{
Fetch individual player data from all matches played
}
\examples{
ElyssePerry <- fetch_player_data(275487,"ODI")
MegLanning <- fetch_player_data(329336,"Test")
SteveSmith <- fetch_player_data(267192,"T20")

}
